# perf: improve serialize bytes