### Enhance Benchmarking and Serialization Logic

 - **`src/ser.rs`**:
   - Optimized `put_slice` method in `MaybeFlip` to handle non-flip cases more efficiently by directly using `put_slice`.
 - **`benches/serde.rs`**:
   - Added a new `bytes` benchmark function to test serialization performance for varying byte sizes.

### Performance

Before:
```
bytes/size-10           time:   [30.636 ns 30.659 ns 30.685 ns]
                        change: [-0.1658% +0.0326% +0.1849%] (p = 0.77 > 0.05)
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild
bytes/size-100          time:   [79.557 ns 80.048 ns 80.754 ns]
                        change: [+0.8302% +1.9252% +3.6712%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
bytes/size-1000         time:   [456.90 ns 458.61 ns 460.58 ns]
                        change: [+0.2209% +0.4600% +0.7241%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
```

After:

```
bytes/size-10           time:   [29.912 ns 29.941 ns 29.972 ns]
                        change: [-2.3758% -2.2718% -2.1584%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
bytes/size-100          time:   [60.175 ns 60.233 ns 60.292 ns]
                        change: [-26.505% -25.471% -24.661%] (p = 0.00 < 0.05)
                        Performance has improved.
bytes/size-1000         time:   [262.92 ns 263.11 ns 263.32 ns]
                        change: [-42.499% -42.343% -42.203%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```